+++ cmd/root.go
@@ -171,6 +171,7 @@ func init() {
 	rootCmd.AddCommand(playMusicCmd)
 	rootCmd.AddCommand(auditCmd)
 	rootCmd.AddCommand(initCmd)
+	rootCmd.AddCommand(upgradeCmd)
 }
 
 func SetVersion(v string) {
@@ -191,7 +192,7 @@ func Execute() {
 	knownCmds := map[string]bool{
 		"install": true, "i": true, "add": true, "run": true, "remove": true, "clean": true,
 		"help": true, "completion": true, "version": true,
-		"_play-sound": true, "_play-music": true, "audit": true, "init": true,
+		"_play-sound": true, "_play-music": true, "audit": true, "init": true, "upgrade": true,
 	}
 
 	if scriptName := firstNonFlagArg(os.Args[1:]); scriptName != "" && !knownCmds[scriptName] {